pallet_revive::evm

Trait GasEncoder

pub trait GasEncoder<Balance>: Sealed {
    // Required methods
    fn encode(gas_limit: U256, weight: Weight, deposit: Balance) -> U256;
    fn decode(gas: U256) -> Option<(Weight, Balance)>;

    // Provided method
    fn as_encoded_values(weight: Weight, deposit: Balance) -> (Weight, Balance) { ... }
}

Encodes/Decodes EVM gas values.

Note

This is defined as a trait rather than standalone functions to allow it to be added as an associated type to crate::Config. This way, it can be invoked without requiring the implementation bounds to be explicitly specified.

This trait is sealed and cannot be implemented by downstream crates.

Required Methods

fn encode(gas_limit: U256, weight: Weight, deposit: Balance) -> U256

Encodes all components (deposit limit, weight reference time, and proof size) into a single gas value.

fn decode(gas: U256) -> Option<(Weight, Balance)>

Decodes the weight and deposit from the encoded gas value. Returns None if the gas value is invalid

Provided Methods

fn as_encoded_values(weight: Weight, deposit: Balance) -> (Weight, Balance)

Returns the encoded values of the specified weight and deposit.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.