pallet_revive::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

call

Makes a call to an account, optionally transferring some balance.

dispatch_as_fallback_account

Dispatch an call with the origin set to the callers fallback address.

eth_transact

A raw EVM transaction, typically dispatched by an Ethereum JSON-RPC server.

instantiate

Instantiates a contract from a previously deployed wasm binary.

instantiate_with_code

Instantiates a new contract from the supplied code optionally transferring some balance.

map_account

Register the callers account id so that it can be used in contract interactions.

remove_code

Remove the code stored under code_hash and refund the deposit to its owner.

set_code

Privileged function that changes the code of an existing contract.

unmap_account

Unregister the callers account id in order to free the deposit.

upload_code

Upload new code without instantiating a contract from it.