pallet_revive::pallet::dispatchables

Function dispatch_as_fallback_account

pub fn dispatch_as_fallback_account<T: Config>(
    call: Box<<T as Config>::RuntimeCall>,
)

Dispatch an call with the origin set to the callers fallback address.

Every AccountId32 can control its corresponding fallback account. The fallback account is the AccountId20 with the last 12 bytes set to 0xEE. This is essentially a recovery function in case an AccountId20 was used without creating a mapping first.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::dispatch_as_fallback_account.