pallet_revive::pallet::dispatchables

Function unmap_account

pub fn unmap_account<T: Config>()

Unregister the callers account id in order to free the deposit.

There is no reason to ever call this function other than freeing up the deposit. This is only useful when the account should no longer be used.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::unmap_account.