pallet_revive::tracing

Function trace

pub fn trace<R, F: FnOnce() -> R>(
    tracer: &mut (dyn Tracer + 'static),
    f: F,
) -> R

Trace the execution of the given closure.

Warning

Only meant to be called from off-chain code as its additional resource usage is not accounted for in the weights or memory envelope.