Module pallet_revive::evm

Types, and traits to integrate pallet-revive with EVM.

Re-exports

pub use rlp;

Modules

runtime

Runtime types for integrating pallet-revive with the EVM.

Macros

if_ethbloom

Conditional compilation depending on whether ethereum-types is built with ethbloom support.

Structs

AccessListEntry

Access list entry

Account

A simple account that can sign transactions

Block

Block object

Bloom

Bloom hash type with 256 bytes (2048 bits) size.

BloomRef

Byte

u8 wrapper type for encoding and decoding hex strings

Bytes

Vec<u8> wrapper type for encoding and decoding hex strings

Bytes8

[u8; 8] wrapper type for encoding and decoding hex strings

Bytes256

[u8; 256] wrapper type for encoding and decoding hex strings

FromStrRadixErr

The error type for parsing numbers from strings.

GenericTransaction

Transaction object generic to all types

H32

H64

H128

Fixed-size uninterpreted hash type with 16 bytes (128 bits) size.

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H264

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

H520

Log

log

ReceiptInfo

Receipt information

SyncingProgress

Syncing progress

Transaction1559Signed

Signed 1559 Transaction

Transaction1559Unsigned

EIP-1559 transaction.

Transaction2930Signed

Signed 2930 Transaction

Transaction2930Unsigned

EIP-2930 transaction.

Transaction4844Signed

Signed 4844 Transaction

Transaction4844Unsigned

EIP-4844 transaction.

TransactionInfo

Transaction information

TransactionLegacySigned

Signed Legacy Transaction

TransactionLegacyUnsigned

Legacy transaction.

Type0

Transaction type identifier: 0

Type1

Transaction type identifier: 1

Type2

Transaction type identifier: 2

U64

Little-endian large integer type Unsigned 64-bit integer.

U128

Little-endian large integer type 128-bit unsigned integer.

U256

Little-endian large integer type 256-bit unsigned integer.

U512

Little-endian large integer type 512-bits unsigned integer.

Withdrawal

Validator withdrawal

Enums

BlockNumberOrTag

Block number or tag

BlockNumberOrTagOrHash

Block number, tag, or block hash

BlockTag

Block tag earliest: The lowest numbered block the client has available; finalized: The most recent crypto-economically secure block, cannot be re-orged outside of manual intervention driven by community coordination; safe: The most recent block that is safe from re-orgs under honest majority and certain synchronicity assumptions; latest: The most recent block in the canonical chain observed by the client, this block may be re-orged out of the canonical chain even under healthy/normal conditions; pending: A sample next block built by the client on top of latest and containing the set of transactions usually taken from local mempool. Before the merge transition is finalized, any call querying for finalized or safe block MUST be responded to with -39001: Unknown block error

BloomInput

FromDecStrErr

Conversion from decimal string error

FromStrRadixErrKind

A list of error categories encountered when parsing numbers.

H256OrTransactionInfo

SyncingStatus

Syncing status

TransactionSigned

TransactionUnsigned

Traits

BigEndianHash

Type Aliases

AccessList

Access list

Address

Public

Secret

Signature