Function pallet_revive::pallet::dispatchables::instantiate

pub fn instantiate<T: Config>(
    value: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    gas_limit: Weight,
    storage_deposit_limit: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    code_hash: H256,
    data: Vec<u8>,
    salt: Option<[u8; 32]>,
)

Instantiates a contract from a previously deployed wasm binary.

This function is identical to [Self::instantiate_with_code] but without the code deployment step. Instead, the code_hash of an on-chain deployed wasm binary must be supplied.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::instantiate.