pallet_revive::pallet

Trait DefaultConfig

pub trait DefaultConfig: DefaultConfig {

    type Time: Time;
    type RuntimeEvent;
    type RuntimeCall;
    type RuntimeHoldReason;
    type WeightPrice;
    type WeightInfo: WeightInfo;
    type Precompiles;
    type FindAuthor: FindAuthor<Self::AccountId>;
    type DepositPerByte;
    type DepositPerItem;
    type CodeHashLockupDepositPercent: Get<Perbill>;
    type UnsafeUnstableInterface: Get<bool>;
    type AllowEVMBytecode: Get<bool>;
    type UploadOrigin;
    type InstantiateOrigin;
    type RuntimeMemory: Get<u32>;
    type PVFMemory: Get<u32>;
    type ChainId: Get<u64>;
    type NativeToEthRatio: Get<u32>;
    type EthGasEncoder;

}

Based on Config. Auto-generated by #[pallet::config(with_default)]. Can be used in tandem with #[register_default_config] and #[derive_impl] to derive test config traits based on existing pallet config traits in a safe and developer-friendly way.

See here for more information and caveats about the auto-generated DefaultConfig trait and how it is generated.

Required Associated Types

type Time: Time

The time implementation used to supply timestamps to contracts through seal_now.

type RuntimeEvent

The overarching event type.

type RuntimeCall

The overarching call type.

type RuntimeHoldReason

Overarching hold reason.

type WeightPrice

Used to answer contracts’ queries regarding the current weight price. This is not used to calculate the actual fee and is only for informational purposes.

type WeightInfo: WeightInfo

Describes the weights of the dispatchables of this module and is also used to construct a default cost schedule.

type Precompiles

Type that allows the runtime authors to add new host functions for a contract to call.

Pass in a tuple of types that implement precompiles::Precompile.

type FindAuthor: FindAuthor<Self::AccountId>

Find the author of the current block.

type DepositPerByte

The amount of balance a caller has to pay for each byte of storage.

Note

It is safe to change this value on a live chain as all refunds are pro rata.

type DepositPerItem

The amount of balance a caller has to pay for each storage item.

Note

It is safe to change this value on a live chain as all refunds are pro rata.

type CodeHashLockupDepositPercent: Get<Perbill>

The percentage of the storage deposit that should be held for using a code hash. Instantiating a contract, protects the code from being removed. In order to prevent abuse these actions are protected with a percentage of the code deposit.

type UnsafeUnstableInterface: Get<bool>

Make contract callable functions marked as #[unstable] available.

Contracts that use #[unstable] functions won’t be able to be uploaded unless this is set to true. This is only meant for testnets and dev nodes in order to experiment with new features.

Warning

Do not set to true on productions chains.

type AllowEVMBytecode: Get<bool>

Allow EVM bytecode to be uploaded and instantiated.

type UploadOrigin

Origin allowed to upload code.

By default, it is safe to set this to EnsureSigned, allowing anyone to upload contract code.

type InstantiateOrigin

Origin allowed to instantiate code.

Note

This is not enforced when a contract instantiates another contract. The Self::UploadOrigin should make sure that no code is deployed that does unwanted instantiations.

By default, it is safe to set this to EnsureSigned, allowing anyone to instantiate contract code.

type RuntimeMemory: Get<u32>

The amount of memory in bytes that parachain nodes a lot to the runtime.

This is used in Pallet::integrity_test to make sure that the runtime has enough memory to support this pallet if set to the correct value.

type PVFMemory: Get<u32>

The amount of memory in bytes that relay chain validators a lot to the PoV.

This is used in Pallet::integrity_test to make sure that the runtime has enough memory to support this pallet if set to the correct value.

This value is usually higher than Self::RuntimeMemory to account for the fact that validators have to hold all storage items in PvF memory.

type ChainId: Get<u64>

The EIP-155 chain ID.

This is a unique identifier assigned to each blockchain network, preventing replay attacks.

type NativeToEthRatio: Get<u32>

The ratio between the decimal representation of the native token and the ETH token.

type EthGasEncoder

Encode and decode Ethereum gas values. Only valid value is (). See GasEncoder.